\name{antialiasReplace}
\alias{qantialias<-}
\alias{antialiasReplace}
\title{Antialiasing}
\usage{qantialias(x) <- value}
\description{Enables or disables antialiasing}
\author{Michael Lawrence}
\arguments{\item{x}{The paint context}
\item{value}{A logical indicating whether antialiasing is enabled}
}

