\name{dashReplace}
\alias{qdash<-}
\alias{dashReplace}
\title{Dash patterns}
\usage{qdash(x) <- value}
\description{Sets the dash pattern}
\author{Michael Lawrence}
\arguments{\item{x}{The paint context}
\item{value}{A numeric vector indicating the pattern; each element
is the length of the corresponding segment}
}

