\name{glyphExpansionReplace}
\alias{qglyphExpansion<-}
\alias{glyphExpansionReplace}
\title{Dash patterns}
\usage{qglyphExpansion(x) <- value}
\description{Sets the glyph expansion, equivalent to \code{cex} in base R}
\author{Michael Lawrence}
\arguments{\item{x}{The paint context}
\item{value}{Floating point multiplier of the glyph size}
}

