\name{qglyphSegment}
\alias{qglyphSegment}
\title{Create a segment glyph with specified length and slope}
\usage{qglyphSegment(x=5, b=0)
}
\description{Create a segment glyph with specified length and slope}
\details{Create a segment glyph with specified length and slope}
\value{a glyph}
\author{Yihui Xie <\url{http://yihui.name}>}
\arguments{\item{x}{length}
\item{b}{slope}
}

