\name{stroke-fill}
\alias{qhasStroke<-}
\title{Stroke and fill colors}
\usage{qhasStroke(x) <- value
       qhasFill(x) <- value
       qstrokeColor(x) <- value
       qfillColor(x) <- value}
\description{Functions for controlling the stroke and fill colors, including
the ability to disable or enable stroking and filling.}
\value{\code{qhasStroke<-}: The stroke or fill color, or a logical value for
\code{qHasStroke} and \code{qHasFill}.

}
\author{Michael Lawrence}
\arguments{\item{x}{The paint context}
\item{value}{The stroke or fill color, or a logical value for
\code{qHasStroke<-} and \code{qHasFill<-}.}
}
\alias{qhasFill<-}
\alias{qstrokeColor<-}
\alias{qfillColor<-}

